+++ hal/lib/athena/DigitalInternal.h
@@ -45,11 +45,11 @@ constexpr int32_t kExpectedLoopTiming = 40;
  * scaling is implemented as an output squelch to get longer periods for old
  * devices.
  */
-constexpr float kDefaultPwmPeriod = 5.05;
+constexpr double kDefaultPwmPeriod = 5.05;
 /**
  * kDefaultPwmCenter is the PWM range center in ms
  */
-constexpr float kDefaultPwmCenter = 1.5;
+constexpr double kDefaultPwmCenter = 1.5;
 /**
  * kDefaultPWMStepsDown is the number of PWM steps below the centerpoint
  */
